Refactor to move semsimian requesting to a service class from config, started a definition of a multi-compare endpoint
